api/requirements/base.txt
boto==2.49.0
click==6.7
Django==1.11.26
django-redis==4.10.0
django-rest-swagger==2.2.0
django-rq==1.3.1
django-storages==1.8
djangorestframework==3.8.2
django-environ==0.4.5
django-ckeditor==5.8.0
django-nested-admin==3.2.4
django-suit==0.2.28
django-cors-headers==2.3.0
django-oauth-toolkit==1.1.3
isoweek==1.3.3
Markdown==2.6.11
mysqlclient==1.4.6
Pillow==5.4.1
python-monkey-business==1.0.0
PyYAML==5.2
redis==3.3.11
rq==0.13
rq-scheduler==0.8.3
six==1.13.0
requests==2.22.0
python-dateutil==2.7.3
coreapi==2.3.3
python-magic==0.4.15
opencc-python-reimplemented==0.1.4
urllib3==1.23
sentry-sdk==0.18.0